FossilOrigin-Name: 6f3640a71d41419b8342cf354ae0586a8abe11f4683a2d3c96be2a3cb177cc24

+++ src/cache.c
@@ -51,7 +51,7 @@ static sqlite3 *cacheOpen(int bForce){
   zDbName = cacheName();
   if( zDbName==0 ) return 0;
   if( bForce==0 ){
-    sz = file_size(zDbName);
+    sz = file_size(zDbName, ExtFILE);
     if( sz<=0 ){
       fossil_free(zDbName);
       return 0;
@@ -63,23 +63,27 @@ static sqlite3 *cacheOpen(int bForce){
     sqlite3_close(db);
     return 0;
   }
-  rc = sqlite3_exec(db,
-     "PRAGMA page_size=8192;"
-     "CREATE TABLE IF NOT EXISTS blob(id INTEGER PRIMARY KEY, data BLOB);"
-     "CREATE TABLE IF NOT EXISTS cache("
-       "key TEXT PRIMARY KEY,"     /* Key used to access the cache */
-       "id INT REFERENCES blob,"   /* The cache content */
-       "sz INT,"                   /* Size of content in bytes */
-       "tm INT,"                   /* Last access time (unix timestampe) */
-       "nref INT"                  /* Number of uses */
-     ");"
-     "CREATE TRIGGER IF NOT EXISTS cacheDel AFTER DELETE ON cache BEGIN"
-     "  DELETE FROM blob WHERE id=OLD.id;"
-     "END;",
-     0, 0, 0);
-  if( rc!=SQLITE_OK ){
-    sqlite3_close(db);
-    return 0;
+  sqlite3_busy_timeout(db, 5000);
+  if( sqlite3_table_column_metadata(db,0,"blob","key",0,0,0,0,0)!=SQLITE_OK ){
+    rc = sqlite3_exec(db,
+       "PRAGMA page_size=8192;"
+       "CREATE TABLE IF NOT EXISTS blob(id INTEGER PRIMARY KEY, data BLOB);"
+       "CREATE TABLE IF NOT EXISTS cache("
+         "key TEXT PRIMARY KEY,"     /* Key used to access the cache */
+         "id INT REFERENCES blob,"   /* The cache content */
+         "sz INT,"                   /* Size of content in bytes */
+         "tm INT,"                   /* Last access time (unix timestampe) */
+         "nref INT"                  /* Number of uses */
+       ");"
+       "CREATE TRIGGER IF NOT EXISTS cacheDel AFTER DELETE ON cache BEGIN"
+       "  DELETE FROM blob WHERE id=OLD.id;"
+       "END;",
+       0, 0, 0
+    );
+    if( rc!=SQLITE_OK ){
+      sqlite3_close(db);
+      return 0;
+    }
   }
   return db;
 }
@@ -167,13 +171,21 @@ void cache_write(Blob *pContent, const char *zKey){
   rc = sqlite3_changes(db);
 
   /* If the write was successful, truncate the cache to keep at most
-  ** max-cache-entry entries in the cache */
+  ** max-cache-entry entries in the cache.
+  **
+  ** The cache entry replacement algorithm is approximately LRU
+  ** (least recently used).  However, each access of an entry buys
+  ** that entry an extra hour of grace, so that more commonly accessed
+  ** entries are held in cache longer.  The extra "grace" allotted to
+  ** an entry is limited to 2 days worth.
+  */
   if( rc ){
     nKeep = db_get_int("max-cache-entry",10);
     sqlite3_finalize(pStmt);
     pStmt = cacheStmt(db,
                  "DELETE FROM cache WHERE rowid IN ("
-                    "SELECT rowid FROM cache ORDER BY tm DESC"
+                    "SELECT rowid FROM cache"
+                    " ORDER BY (tm + 3600*min(nRef,48)) DESC"
                     " LIMIT -1 OFFSET ?1)");
     if( pStmt ){
       sqlite3_bind_int(pStmt, 1, nKeep);
@@ -186,6 +198,13 @@ cache_write_end:
   sqlite3_exec(db, rc ? "COMMIT" : "ROLLBACK", 0, 0, 0);
   sqlite3_close(db);
 }
+
+/*
+** SETTING: max-cache-entry                 width=10 default=10
+**
+** This is the maximum number of entries to allow in the web-cache
+** for tarballs, ZIP-archives, and SQL-archives.
+*/
 
 /*
 ** Attempt to read content out of the cache with the given zKey.  Return
@@ -252,6 +271,8 @@ void cache_initialize(void){
 **    list|ls      List the keys and content sizes and other stats for
 **                 all entries currently in the cache.
 **
+**    size ?N?     Query or set the maximum number of entries in the cache.
+**
 **    status       Show a summary of the cache status.
 **
 ** The cache is stored in a file that is distinct from the repository
@@ -293,7 +314,10 @@ void cache_cmd(void){
     }else{
       fossil_print("nothing to clear; cache does not exist\n");
     }
-  }else if(( strncmp(zCmd, "list", nCmd)==0 ) || ( strncmp(zCmd, "ls", nCmd)==0 )){
+  }else if( strncmp(zCmd, "list", nCmd)==0
+        ||  strncmp(zCmd, "ls", nCmd)==0 
+        ||  strncmp(zCmd, "status", nCmd)==0
+  ){
     db = cacheOpen(0);
     if( db==0 ){
       fossil_print("cache does not exist\n");
@@ -308,32 +332,73 @@ void cache_cmd(void){
       );
       if( pStmt ){
         while( sqlite3_step(pStmt)==SQLITE_ROW ){
-          fossil_print("%s %4d %8s %s\n",
-             sqlite3_column_text(pStmt, 3),
-             sqlite3_column_int(pStmt, 2),
-             sqlite3_column_text(pStmt, 1),
-             sqlite3_column_text(pStmt, 0));
+          if( zCmd[0]=='l' ){
+            fossil_print("%s %4d %8s %s\n",
+               sqlite3_column_text(pStmt, 3),
+               sqlite3_column_int(pStmt, 2),
+               sqlite3_column_text(pStmt, 1),
+               sqlite3_column_text(pStmt, 0));
+          }
           nEntry++;
         }
         sqlite3_finalize(pStmt);
       }
       sqlite3_close(db);
-      fossil_print("Entries: %d  Cache-file Size: %lld\n",
-                   nEntry, file_size(zDbName));
+      fossil_print(
+         "Filename:        %s\n"
+         "Entries:         %d\n"
+         "max-cache-entry: %d\n"
+         "Cache-file Size: %,lld\n",
+         zDbName,
+         nEntry,
+         db_get_int("max-cache-entry",10),
+         file_size(zDbName, ExtFILE)
+      );
       fossil_free(zDbName);
     }
-  }else if( strncmp(zCmd, "status", nCmd)==0 ){
-    fossil_print("TBD...\n");
+  }else if( strncmp(zCmd, "size", nCmd)==0 ){
+    if( g.argc>=4 ){
+      int n = atoi(g.argv[3]);
+      if( n>=5 ) db_set_int("max-cache-entry",n,0);
+    }
+    fossil_print("max-cache-entry: %d\n", db_get_int("max-cache-entry",10));
   }else{
     fossil_fatal("Unknown subcommand \"%s\"."
-                 " Should be one of: clear init list status", zCmd);
+                 " Should be one of: clear init list size status", zCmd);
   }
 }
 
 /*
+** Given a cache key, find the check-in hash and return it as a separate
+** string.  The returned string is obtained from fossil_malloc() and must
+** be freed by the caller.
+**
+** Return NULL if not found.
+**
+** The key is usually in a format like these:
+**
+**    /tarball/HASH/NAME
+**    /zip/HASH/NAME
+**    /sqlar/HASH/NAME
+*/
+static char *cache_hash_of_key(const char *zKey){
+  int i;
+  if( zKey==0 ) return 0;
+  if( zKey[0]!='/' ) return 0;
+  zKey++;
+  while( zKey[0] && zKey[0]!='/' ) zKey++;
+  if( zKey[0]==0 ) return 0;
+  zKey++;
+  for(i=0; zKey[i] && zKey[i]!='/'; i++){}
+  if( !validate16(zKey, i) ) return 0;
+  return fossil_strndup(zKey, i);
+}
+
+
+/*
 ** WEBPAGE: cachestat
 **
-** Show information about the webpage cache.  Requires Admin privilege.
+** Show information about the webpage cache.  Requires Setup privilege.
 */
 void cache_page(void){
   sqlite3 *db;
@@ -342,6 +407,7 @@ void cache_page(void){
 
   login_check_credentials();
   if( !g.perm.Setup ){ login_needed(0); return; }
+  style_set_current_feature("cache");
   style_header("Web Cache Status");
   db = cacheOpen(0);
   if( db==0 ){
@@ -350,30 +416,44 @@ void cache_page(void){
     char *zDbName = cacheName();
     cache_register_sizename(db);
     pStmt = cacheStmt(db,
-         "SELECT key, sizename(sz), nRef, datetime(tm,'unixepoch')"
+         "SELECT key, sz, nRef, datetime(tm,'unixepoch')"
          "  FROM cache"
-         " ORDER BY tm DESC"
+         " ORDER BY (tm + 3600*min(nRef,48)) DESC"
     );
     if( pStmt ){
       @ <ol>
       while( sqlite3_step(pStmt)==SQLITE_ROW ){
         const unsigned char *zName = sqlite3_column_text(pStmt,0);
+        char *zHash = cache_hash_of_key((const char*)zName);
         @ <li><p>%z(href("%R/cacheget?key=%T",zName))%h(zName)</a><br />
-        @ size: %s(sqlite3_column_text(pStmt,1))
+        @ size: %,lld(sqlite3_column_int64(pStmt,1))
         @ hit-count: %d(sqlite3_column_int(pStmt,2))
-        @ last-access: %s(sqlite3_column_text(pStmt,3))</p></li>
+        @ last-access: %s(sqlite3_column_text(pStmt,3)) \
+        if( zHash ){
+          @ %z(href("%R/timeline?c=%S",zHash))check-in</a>\
+          fossil_free(zHash);
+        }
+        @ </p></li>
+        
       }
       sqlite3_finalize(pStmt);
       @ </ol>
     }
     zDbName = cacheName();
-    bigSizeName(sizeof(zBuf), zBuf, file_size(zDbName));
-    @ <p>cache-file name: %h(zDbName)</p>
-    @ <p>cache-file size: %s(zBuf)</p>
+    bigSizeName(sizeof(zBuf), zBuf, file_size(zDbName, ExtFILE));
+    @ <p>
+    @ cache-file name: %h(zDbName)<br>
+    @ cache-file size: %s(zBuf)<br>
+    @ max-cache-entry: %d(db_get_int("max-cache-entry",10))
+    @ </p>
+    @ <p>
+    @ Use the "<a href="%R/help?cmd=cache">fossil cache</a>" command
+    @ on the command-line to create and configure the web-cache.
+    @ </p>
     fossil_free(zDbName);
     sqlite3_close(db);
   }
-  style_footer();
+  style_finish_page();
 }
 
 /*
@@ -394,9 +474,10 @@ void cache_getpage(void){
   zKey = PD("key","");
   blob_zero(&content);
   if( cache_read(&content, zKey)==0 ){
+    style_set_current_feature("cache");
     style_header("Cache Download Error");
     @ The cache does not contain any entry with this key: "%h(zKey)"
-    style_footer();
+    style_finish_page();
     return;
   }
   cgi_set_content(&content);
